requirements.txt
google-genai
line-bot-sdk<3.0.0
tenacity
flask
